+++ scripts/src/ui.js
@@ -423,7 +423,16 @@
 
                     <div class="ras-field">
                         <label style="display:inline-flex; align-items:center; margin-right: 15px;">
+                            <input type="checkbox" id="ras-tag-broken" ${STATE.config.tagBrokenLinks ? 'checked' : ''} style="margin-right:5px;"> Tag Broken Links
+                        </label>
+                    </div>
+
+                    <div class="ras-field">
+                        <label style="display:inline-flex; align-items:center; margin-right: 15px;">
                              <input type="checkbox" id="ras-delete-empty" ${STATE.config.deleteEmptyCols ? 'checked' : ''} style="margin-right:5px;"> Delete Empty Folders
+                        </label>
+                        <label style="display:inline-flex; align-items:center;">
+                             <input type="checkbox" id="ras-nested-collections" ${STATE.config.nestedCollections ? 'checked' : ''} style="margin-right:5px;"> Allow Nested Folders
                         </label>
                     </div>